Create Ambrosio's Grocery Store
Ambrosio is a friend of the neighborhood and he decided to give discounts to please his customers. Ambrósio's grocery store contains only 04 items, whose prices are given by the table below:

1 - 5,30
2 - 6,00
3 - 3,20
4 - 2,50

The discount rule is quite simple: if the quantity of products purchased is equal to or greater than fifteen, then the discount is granted. The discount is also granted if the total purchase value is greater than or equal to 40 reais (Brazilian currency).

The discount's value is 15%.

Your mission is to make a program that reads the product code, the quantity purchased and prints the amount that the customer must pay, considering the discount, when applicable.

Consider that the customer can only purchase a single type of product each time they use your software.