Create Insurance Criteria
An insurance broker charges cheaper if the main driver of the vehicle is a woman and if he/she is older than 40. Otherwise, the insurance becomes expensive

Write a program that reads an integer (0 or 1) which indicates if the main driver is a man and another integer (0 or 1) which represents if the main driver is older than 40. Based on the read values, the program must indicate, using 0 or 1, if the insurance will be cheap (true - 1) or expensive (false - 0)