Create Triangles Classification
The simplest triangles are classified according to the limit of the relative proportions of their sides:

An equilateral triangle is a triangle with all three sides of equal length, in other words, all 3 sides are congruent. A triangle is also equilateral if all of its intern angles are congruent (measuring 60°), and therefore, it is classified as a regular polygon. 

An isosceles triangle has at least two sides of the same length (the legs) and two congruent angles. An equilateral triangle is, consequently, a special case of isosceles triangle, where not 2, but all 3 sides are equal, as well as the angles, which measure 60°. In an isosceles triangle, the angle formed by the two congruent sides is named vertex angle. The other angles are called base angles and they are congruent.

In a scalene triangle, all three sides are different. The internal angles of a scalene triangle also have different measures. 

Your mission is write a program to classify a triangle whose sides will be given, as: scalene (the three sides' lengths are different), isosceles (two sides with the same length) or equilateral (the three sides have the same length).