Create Water bill
The local water supply company, SBC, is promoting a water conservation campaign, distributing leaflets and promoting actions demonstrating the importance of water for life and the environment.

To further encourage water savings, SBC has altered its supply prices so that the customers who consume less water will, proportionally, pay less for each m³.

Every customer pays a monthly subscription of $ 7, which includes a franchise of 10 m³ of water. That is, for any consumption between 0and 10 m³, the consumer pays the same amount of $ 7 (note that the subscription fee must be paid even if the consumer has not consumed water). Above 10 m³,each subsequent cubic meter has a different value, depending on the consumption. SBC charges only for whole quantities of consumed cubic meters .

The table below specifies the price per m³ for each consumption range:
Thus, for example, if the consumption was 120 m³, the value of the bill is:

$ 7 of the basic subscription;
$ 20 for consumption in the range 11-30 m³;
$ 140 for consumption in the range 31 - 100 m³;
$ 100 for consumption in the range 101 - 120 m³.
Then the total amount of the water bill is $ 267. 

Implement a program that, given the consumption of a residence in m³, calculates the value of the water bill of that residence.