Create Slow down, slow down, watch out for the speed radar!
The Getran (Education Group for Traffic) decided to install a radar on the way between your home and your college in order to limit the speed of the cars on a certain road.

But They decided to innovate and the allowed speed for the road can be changed daily. For example, on holidays, They usually put the speed limit at 80Km/h. But during the week, it is common to reduce the speed limit to 60km/h.

For this, digital plates informs the drivers about the current speed limit.

The table of penalties for those who exceed those limits is:

Speed up to 20% higher than allowed: fine of R$85.13 and 4 points in the driver’s license;
Speed greater than 20% and up to 50% higher than the allowed: fine of R$ 127.69 and 5 points in the driver’s license;
Speed 50% higher than the allowed: fine of R$ 574.62, 7 points in the driver’s license, seizure of the driver’s license and suspension of the right to drive.
You will receive the speed limit of the road and the car’s speed and you must calculate the amount of the fine and how many points the driver lost in the driver’s license, if he has exceeded the limit.