Create Cinema Ticket
Normally, the price is R$ 20,00 (Brazilian Currency). However, if the client is a student or an elderly person, the price for the tickets is halved

Write a program that reads a integer (0 or 1) which indicates if the client is a student and then another integer (0 or 1) which represents whether the client is an elderly person or not. Based on the read values, the program must indicate, using 0 or 1, if the client pays the full price (false - 0) or if the price is halved (true - 1).